+++ FlutterPlugin/ios/Classes/Native/DaroNativeAdFactory.swift
@@ -2,5 +2,8 @@ import UIKit
 import Daro
 
 public protocol DaroNativeAdFactory {
-    func createNativeAdView(unit: DaroAdUnit) -> DaroAdNativeView
+    func createNativeAdView(
+        unit: DaroAdUnit,
+        preferredAdChoicesPosition: DaroAdChoicesPosition
+    ) -> DaroAdNativeView
 }