+++ FlutterPlugin/lib/src/native_ad/daro_ad_choices_position.dart
@@ -1,0 +1,6 @@
+enum DaroAdChoicesPosition {
+  topLeft,
+  topRight,
+  bottomRight,
+  bottomLeft,
+}